fix: satisfy pre-push gates for the ephemeral fine-tune and tunnel work
